+++ R/app_ui.R
@@ -17,10 +17,8 @@ app_ui <- function(request) {
         theme = "auto",
         dark = FALSE,
         filled = TRUE,
-        pullToRefresh = TRUE
+        pullToRefresh = FALSE
       ),
-      # PWA compatibility
-      PWAMask(),
       # Waiting background before the UI is ready to be shown
       waiter::waiterPreloader(
         html = waiter::spin_1(),
@@ -55,7 +53,6 @@ app_ui <- function(request) {
           mod_share_ui("share_ui"),
           mod_request_ui("request_ui"),
           mod_about_ui("about_ui"),
-          mod_more_ui("more_ui"),
           mod_approval_ui("approval_ui")
         )
       )