+++ NEWS.md
@@ -1,3 +1,14 @@
+# shareIBC 0.1.2
+* PWA capabilities had been properly setup. Lighthouse tests have been passed successfully for PWA features.
+* A new icon has been designed, an array of icons for different devices has been provided.
+* Instructions for the request of removal of a post have been placed both in the post tabs and in the poster confirmation mail.
+* Posts: The image size limit has been increased to 10MB.
+* UI:
+    - The post button and icon has been improved to make it more clear.
+    - The subscription workflow has been passed to the Share tab. Its logic is placed in a separate script.
+    - The more app has been deleted and its contents have been merged with the about us tab.
+
+
 # shareIBC 0.1.1
 * The visualization error of the table of economic income required for the food program has been corrected.
 * The approval tab can't be accessed unless a valid post id and key is provided.